0.0.3

+++ Classes/YQRatingPicker.h
@@ -12,10 +12,10 @@ NS_ASSUME_NONNULL_BEGIN
 
 @interface YQRatingPicker : UIView
 
-@property(nonatomic) IBInspectable double value;                        // default is 0. sends UIControlEventValueChanged. clamped to min/max
-@property(nonatomic)IBInspectable double minimumValue;                 // default 0. must be less than maximumValue
-@property(nonatomic) IBInspectable double maximumValue;                 // default 100. must be greater than minimumValue
-@property(nonatomic) IBInspectable double stepValue;
+@property(nonatomic) IBInspectable NSInteger value;                        // default is 0. sends UIControlEventValueChanged. clamped to min/max
+@property(nonatomic)IBInspectable NSInteger minimumValue;                 // default 0. must be less than maximumValue
+@property(nonatomic) IBInspectable NSInteger maximumValue;                 // default 100. must be greater than minimumValue
+@property(nonatomic) IBInspectable NSInteger stepValue;
 
 @end
